+++ src/main/java/nz/co/searchwellington/model/TagContentCount.java
@@ -5,17 +5,17 @@ import nz.co.searchwellington.model.frontend.FrontendTag;
 public class TagContentCount {
     
     private FrontendTag tag;
-    private int count;
+    private long count;
     
-    public TagContentCount(FrontendTag tag, int count) {
+    public TagContentCount(FrontendTag tag, long count) {
         this.tag = tag;
         this.count = count;
     }
     
-    public int getCount() {
+    public long getCount() {
         return count;
     }
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
     public FrontendTag getTag() {
